+++ meu-app/backend/src/routes/user.ts
@@ -1,13 +1,129 @@
-// Define os endpoints da API e mapeia URLs para os controladores.
+// src/routes/user.ts
+
 import { Router } from "express";
 import { getUser, createUser, loginUser, deleteUser } from "../controllers/userController";
 
 const router = Router();
 
-// Definir as rotas e associá-las aos controladores
+/**
+ * @swagger
+ * tags:
+ *   name: Users
+ *   description: Operações relacionadas a usuários
+ */
+
+/**
+ * @swagger
+ * /users/{id}:
+ *   get:
+ *     summary: Obter um usuário pelo ID
+ *     tags: [Users]
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         schema:
+ *           type: integer
+ *         required: true
+ *         description: ID do usuário
+ *     responses:
+ *       200:
+ *         description: Dados do usuário
+ *       404:
+ *         description: Usuário não encontrado
+ */
 router.get("/:id", getUser);
+
+/**
+ * @swagger
+ * /users:
+ *   post:
+ *     summary: Criar um novo usuário
+ *     tags: [Users]
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             required:
+ *               - nome
+ *               - sobrenome
+ *               - email
+ *               - cpf_number
+ *               - senha
+ *               - confirmar_senha
+ *             properties:
+ *               nome:
+ *                 type: string
+ *               sobrenome:
+ *                 type: string
+ *               email:
+ *                 type: string
+ *               cpf_number:
+ *                 type: string
+ *               senha:
+ *                 type: string
+ *               confirmar_senha:
+ *                 type: string
+ *     responses:
+ *       201:
+ *         description: Usuário criado com sucesso
+ *       400:
+ *         description: Erro de validação
+ *       409:
+ *         description: Email já está em uso
+ */
 router.post("/", createUser);
+
+/**
+ * @swagger
+ * /users/login:
+ *   post:
+ *     summary: Login de usuário
+ *     tags: [Users]
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             required:
+ *               - email
+ *               - senha
+ *             properties:
+ *               email:
+ *                 type: string
+ *               senha:
+ *                 type: string
+ *     responses:
+ *       200:
+ *         description: Login realizado com sucesso
+ *       400:
+ *         description: Erro de validação
+ *       401:
+ *         description: Credenciais inválidas
+ */
 router.post("/login", loginUser);
+
+/**
+ * @swagger
+ * /users/{id}:
+ *   delete:
+ *     summary: Deletar um usuário pelo ID
+ *     tags: [Users]
+ *     parameters:
+ *       - in: path
+ *         name: id
+ *         schema:
+ *           type: integer
+ *         required: true
+ *         description: ID do usuário
+ *     responses:
+ *       200:
+ *         description: Usuário deletado com sucesso
+ *       404:
+ *         description: Usuário não encontrado
+ */
 router.delete("/:id", deleteUser);
 
 export default router;
